Move reading of a pivot table definition part to a separate reader.
